Move default api urls to a json structure
