+++ apps/contract-verification/src/app/types/index.ts
@@ -1,4 +1,3 @@
 export * from './ThemeType'
 export * from './SettingsTypes'
 export * from './VerificationTypes'
-export * from './defaults'
