+++ apps/contract-verification/src/app/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './default-settings'
