+++ apps/contract-verification/src/app/views/LookupView.tsx
@@ -1,6 +1,8 @@
 import { useContext, useState } from 'react'
 import { SearchableChainDropdown, ContractAddressInput } from '../components'
-import { LookupResponse, mergeChainSettingsWithDefaults, validConfiguration, VerifierIdentifier, VERIFIERS, type Chain } from '../types'
+import { mergeChainSettingsWithDefaults, validConfiguration } from '../utils'
+import type { LookupResponse, VerifierIdentifier, Chain } from '../types'
+import { VERIFIERS } from '../types'
 import { AppContext } from '../AppContext'
 import { CustomTooltip } from '@remix-ui/helper'
 import { getVerifier } from '../Verifiers'
