+++ apps/contract-verification/src/app/views/SettingsView.tsx
@@ -1,7 +1,7 @@
 import { useContext, useState } from 'react'
 import { SearchableChainDropdown, ConfigInput } from '../components'
 import type { VerifierIdentifier, Chain, VerifierSettings, ContractVerificationSettings } from '../types'
-import { mergeChainSettingsWithDefaults } from '../types'
+import { mergeChainSettingsWithDefaults } from '../utils'
 import { AppContext } from '../AppContext'
 
 export const SettingsView = () => {
