+++ apps/contract-verification/tsconfig.json
@@ -5,6 +5,7 @@
     "allowJs": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true,
     // "strict": true
   },
   "files": [],
